Add notes around encryption key usage
This adds a note on how to pass in the encryption key to fleet-server to connect it to Elasticsearch. This is only temporary until changes were made in Kibana.

Additionaly the fleet-server.dev.yml was added to gitignore. This allows to have a local config file which is not checked in.