Cleanup and refactoring
I was going through the code base to get more familiar with it. Along the way I made some small changes / cleanups that I stumbled over. There are also parts where I added small tests which are more for me to understand the code on what it does. Here is the list:

* Remove savedObjectKey function as it is not needed anymore
 * add config options for monitoring to the default configuration for documentation purpose
* move the creation of the nowString to where it is actually used. Seems it is used only once.
* add test for convertActions to better understand the format on my end
* Add note to missing parts
* update t.Fatal should not be used in test go routine
* add basic token test
* add more entries to the config file about available config options
* remove unused env package
* remove not used `missing.txt` file.
* remove dsl readme and filed https://github.com/elastic/fleet-server/issues/136 instead