Add index refresh call after fetch if the list of list has holes
After discussion with Elasticsearch team it looks like the current global checkpoints
API doesn't not gurantee the list of document without holes on
subsequent index search and there is still a slim chance of this with
our current implementation. Elastisearch team is working on getting the
refresh implemented on thier side, and for now recommended to do the
refresh before fetch if there are holes detected in the result.

Added the check for the holes and here are the steps:
1. Call Global checkpoints = 5
2. Search = 1, 2, 3, 5.
3. Manual refresh
4. Search and get 4,5
5. Return to step 1