Add changelog file to fleet-server
Currently fleet-server does not contain a changelog. With this I am adding the simplest changelog file possible. We need to find a better way on how to handle the changelog but for now I think it is more important to have a changelog then not to have one at all.

Closes https://github.com/elastic/fleet-server/issues/602