Improve PR description template (#591)
One area I think we can improve as a team is on descriptions of pull requests. A pull request description should not only reference an issue that it solves but also be able to stand on its own. It is important to describe how a certain problem is solved. What approach was taken and why. This allows anyone reviewing a PR to build a mental model on what change is in the pull request and if it aligns with the described solution.

There are cases where the PR description takes more time then the actual fix in code. Anyone reading the PR description should understand why the changes was made and what the change contains.

I opened to discuss a change to the template with these two question on the top. The template itself from my perspective is mostly meant as a guideline/helper but it should be up to each engineer to decide what is really needed for this PR.

This PR is made on the fleet-server repository as a first trial but the idea is to use the same template for Beats and other repositories.

(cherry picked from commit 0d1a384c45047e7f75acd57129ea2b0eeb6b343c)