instrumentation: Add initial APM tracing (#898)
Adds support to enable instrumentation via the APM Go agent. New config
options have been added to the `Server` input which could be set up in
the `fleet-server` integration configuration.

The added instrumentation covers the `fleet-server` http server and the
Adds support to enable instrumentation via the APM Go agent. New config
options have been added to the `Server` input which could be set up in
the `fleet-server` integration configuration.

The added instrumentation covers the `fleet-server` http server and the
`go-elasticsearch` client.

A sample of the configuration that's been added (`instrumentation`):

```yaml
inputs:
  - type: fleet-server
    server:
      instrumentation:
        enabled: true
        hosts: ["localhost:8200"]
        environment: production
        secret_token: token
        api_key: apikey
```

Signed-off-by: Marc Lopez Rubio <marc5.12@outlook.com>