Update CODEOWNERS file (#1221)
This fixes the / with * based on the CODEOWNERS docs and removes robots as the owners of the Jenkins file. Many changes to the Jenkinsfiles are likely pushed by the robots team but the owners and reviewers is the control plane team.

(cherry picked from commit e4744b4234bc0dfda78b3141e5151f2bfd246104)

# Conflicts:
#	.github/CODEOWNERS