Correctly fix the `goe` dependency
This add the required overrides and rules for the CC0,
The CC0 is in our green light list.

Notes: This is an indirect dependencies coming from beats, this is
present in both beats and apm-server. Both of theses project doesn't
have rules or overrides in place to allow it. I've lost time to investigate why
the license detector didn't raise any issue in their repository.

Looking at both projects `go.sum` the dependencies is defined.
If you run `go list -m all` you will see it. The output of this command
result is used by our `go-licence-detector` to flag incompatible license.