Always add `api_key` to Elasticsearch output
This fixes a few issues:

- When the hash on the permission and the hash on the agent were the
same the key were not added to the output. This was causing the agent to
not be able to connect to Elasticsearch. If you changes an integration
policy field, like the host,  and the data stream are the same the permission hash will
stay the same.

- Fix a possible issue when the new key was not saved on the Agent model, the
key was only saved when we were targetting a default agent policy.
Currently we only support a single ES output so this should always be saved.

- Reorganize the test to ensure that all the scenario from the case
statement are tested.

Fix: https://github.com/elastic/elastic-agent/issues/285