Fleet-server package cache implementation
Allow the fleet-server to serve packages that the elastic-agent uses for
upgrades. The endpoint for packages may be configured with an associated
concurrency limit and a bandwidth limit. The cache may have an
associated retention duration. When the fleet-server needs to retrieve a
package from the upstream source it will download the associated .sha512
and .asc files in order to verify the package.