Change pkg/monitor/mock to use testify/mock
Use testify/mock in monitor/mock and change test cases that use it.
Using a real mocking lib allows us to be explicit on the exact calls
that we need when testing.
The monitor mock is unable to reside in pkg/testing alongside the
BulkMock as that would introduce a circular dependancy when running
integration tests.