Cleanup the test suite
This remove an issues when multiple goroutine wanted to change the test
logger, this was tripping the race detector. But I haven't seen that in
CI.

This actually encalsulate the access of the policiesCanceller under an
accessor, this aint perfect, because I think the object is actually
leaking internal states with the different executed goroutine.

(cherry picked from commit b56d37bd547c5aa3a588ad6f27a08ccc208076ff)