Remove events from agent checkin body. (#1842)
Remove the events attribute from the agent checkin body. Note that
removal of the attribute will not stop the server from issuing a 400 if
the response body is too long. The removal is so that the checkin code
on the fleet-server and agent remain comparable.

Co-authored-by: Blake Rouse <blake.rouse@elastic.co>
(cherry picked from commit 8a22669ebe8a2ef0af9e05c4ea53992131217e2e)

# Conflicts:
#	CHANGELOG.next.asciidoc