Don't send POLICY_CHANGE actions retrieved from index to agent. (#1963)
* Don't send POLICY_CHANGE actions retrieved from index to agent.

The fleet-server should not send any policy change actions that are
written to the actions index to an agent on checkin. The server will
remove these actions in the convert method and emit a warning message.
The ack token that is used is not altered in this case. Policy change
actions are dynamically generated by the fleet-server when it detects
that the agent is not running an up to date version of the policy.

* move filtering to its own method

* Fix linter, tests, fix file name

(cherry picked from commit b51bf97530f988882ba3886be7919003042a9d56)

# Conflicts:
#	CHANGELOG.next.asciidoc