Add SNAPSHOT only stand alone mode
Allow the fleet-server to run independantly of the elastic-agent when a
SNAPSHOT is built.
When in stand alone mode the fleet-server agent will create an agent
record for itself in the agents index with the ID specified in the
fleet.yml file. After regular startup the agent will also send checkins
to it's API handler in order to appear as a healthy agent in the UI.
All config is read from the fleet.yml file. Actions and policy changes
are ignored. action seq numbers are not persisted.