Allow dev stand-alone execution (#2212)
* Add SNAPSHOT only stand alone mode

Allow the fleet-server to run independantly of the elastic-agent when a
SNAPSHOT is built.
When in stand alone mode the fleet-server agent will create an agent
record for itself in the agents index with the ID specified in the
fleet.yml file. After regular startup the agent will also send checkins
to it's API handler in order to appear as a healthy agent in the UI.
All config is read from the fleet.yml file. Actions and policy changes
are ignored. action seq numbers are not persisted.

* Add license headers

* test fixes

* Change to DEV flag, add readme

* minor cleanup

* add old config for testing

* fix stand alone yml

* Fix data race in tests caused by logger contention

* Fix setup issue

* Add tls instructions

* return default load limits when passed a 0 value

* Update internal/pkg/server/standAlone_release.go

Co-authored-by: Anderson Queiroz <me@andersonq.me>

* Update internal/pkg/server/standAlone_release.go

Co-authored-by: Anderson Queiroz <me@andersonq.me>

* Update README.md

Co-authored-by: Anderson Queiroz <me@andersonq.me>

* Update internal/pkg/server/standAlone_dev.go

Co-authored-by: Anderson Queiroz <me@andersonq.me>

* Update internal/pkg/server/standAlone_dev.go

Co-authored-by: Anderson Queiroz <me@andersonq.me>

* removed commented code

* added changelog

* standalone:false for integration test

Co-authored-by: Julia Bardi <90178898+juliaElastic@users.noreply.github.com>
Co-authored-by: Anderson Queiroz <me@andersonq.me>
Co-authored-by: Julia Bardi <julia.bardi@elastic.co>