Fix duplicate key issue in agent manager
Fix the issue where duplicate server.* keys in the input are being
silently discarded in the agent management layer.
Passing the input soruce as a map directly in one step when creating the
fleet-server config does not merge all keys correctly.
In order to resolve we explicitly create a ucfg.Config from the passed
input and unpack to a map, then use that map when creating the server
config.