Fix benchmark test that required the mock MUpdate
The benchmark test for the benchmarkBulk was missing the mock for the
MUpdate method and thus it was constantly failing with a panic.

```
...
panic:
assert: mock: I don't know what to return because the method call was unexpected.
....
```

This commit adds the required mock.on to make the test pass.

Signed-off-by: Alexandros, Sapranidis <alexandros@elastic.co>