Add poll timeout to checkin requests
Add poll_timeout attribute to checkin request bodies that allow the
client to inform fleet-server of when the request will time out so that
fleet-server can alter the request poll duration and write deadline to
respond before then.
If the attribute is not filled, or has a zero value such as "0s" the
configured timeouts are used instead.