Avoid crashing on startup if Elasticsearch is not available (#2693)
Initial Info request is removed, so the client can be eventually used even if during
startup Elasticsearch is not available.
Subsystems that need initialization keep retrying if they fail.
Standalone monitor keeps running once the service is healthy. If it loses connection
with Elasticsearch, it goes to degraded state.
Degraded state returns 503 error.