Add more span for APM tracking
Add more detailed span for APM tracking of how the fleet-server
processes each endpoint. More attention is paid to any interactions that
the fleet-server has with elasticsearch as currently bulker transactions
do not appear in the http view (we can check if this can be changed).
Also add more spans around any serialization/deserializations as we
don't know what takes time in our responses and want to measure anything
that can be expensive.