Stop using zerolog/log.Logger and use logger from context instead
Stop using the global zerolog logger and use one that is passed through
the context instead. When fleet-server is running normally
context-logger defaults to the same as the global logger, however it is
much easier to change for test cases so we can have log output that's
tied to the specific test making it easier to track any errors in tests.