Cache and limit defaults are derived from system memory.
Allow the fleet-server to get matching limits settings based off system
memory if no agent max is provided. Allow default settings to be used
if a negative value is specified as the agents max setting.
Change ServerLimits and Cache InitDefault functions to nops to fix an
undiscovered bug with loading limits.