Change metrics collection to use OTEL + the APM bridge
Change the underlying metrics collection to use OTEL and the APM bridge
instead of prometheus and the elastic-agent-libs metrics collection
mechanism. All API metrics will now have server.host and server.port
attributes assocaited with the metric so we can easily determine when a
request hits the external or internal port. Provide a translation layer
to expose the OTEL metrics on the /stats endpoint so that Metricbeat can
still collect fleet-server meterics. Note that generic CPU/Mem/System
metrics are no longer collected or exposed on the endpoint.