Add support for bootstrap block in output when running in agent mode
Add support for a bootstrap attribute in the output when running in
agent mode. If this attribute is missing the output block is used
directly. If the attribute is provided, then any attributes within
bootstrap that are not in the parent (output) object are recursivly
injected and the resulting output is tested. If the resulting config
works it is used, if it fails the test the bootstrap config is passed.