Agent mode support for bootstrap output (#3506)
The elastic-agent will inject enrollment configuration options in output.elasticsearch.bootstrap instead of overwriting matching keys in output.elasticsearch.
When running under agent mode, fleet-server will inject specific keys in bootstrap that are not in output.elasticsearch, then test the resulting output to see if it can connect to Elasticsearch, if so it is used. If not then bootstrap is used instead and the output is periodically retested in case the failure was caused by a temporary network issue.