Automatically label PRs to `main` with `backport-v8.x`
When the `main` branch switches over to `9.0`, we will want most PRs that are
merged into `main` to also be backported to `8.x`. This PR makes this process
easier by automatically adding the `backport-8.x` label to all PRs targeting
the `main` branch.

The exceptions will be PRs for changes that should _only_ be present in `9.0`.
For such PRs, authors should simply remove the `backport-8.x` label from their
PR.