Add secret_paths attributes to policies sent to agents (#3908)
Add a new attribute "secret_paths" to the policy data sent to agents.
This attribute is a list of keys where the fleet-server has repalced a
reference with a secret value. The agent is expected to redact the
values of these keys when outputting policy data.

(cherry picked from commit 4864cf419c133ea2b0fda835c4c50b133180a4d8)