Prevent mapping explosion on logs (#4181)
This commit remove some JSON objects that were added to the logs, thus
preventing mapping explosion when those logs are ingested into
Elasticsearch.

Some entries are converted to strings, others are fully removed to keep
the log within a reasonable size and some are kept as string at trace level.