Restore connection-level limiter (#5372)
Restore connection level limiter to prevent OOM incidents. (removed in #4402)
This limiter is used in addition to the request-level throttle so that once
our in-flight requests reaches max_connections a 429 is returned, but if the
total connections the server uses is over max_connections*1.1 the server drops
the connection before the TLS handshake.

(cherry picked from commit 39199eff04229dddbffecd43777f0aa5dd381f56)

# Conflicts:
#	go.mod
#	internal/pkg/api/server.go